Visulization is Currently Finished
Based on Ray marching, the visualization is efficient enough to render no many how large the size of the fluid cube is. But voxel is still kept to have a more stylized rendering result in the future(maybe).
And the basic structure for the stable fluids logic is set up in the "/graphics/fluidcube" file. Just check them.
Hope everything will be understandable. Try my best to comment.